% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{DEFAULT_ENVVARS}
\alias{DEFAULT_ENVVARS}
\title{The default set of environment variables used for all tasks.}
\format{
An object of class \code{hipercow_envvars} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
DEFAULT_ENVVARS
}
\description{
These are chosen to provide the best general experience. The default
variables may be overriden globally by setting the hipercow.default_envvars
option, or on a per-task and variable basis by assigning a different value
to the environment variable in question.
}
\keyword{datasets}
