% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dide.R
\name{dide_check}
\alias{dide_check}
\title{Check we can use the DIDE cluster}
\usage{
dide_check(path = getwd())
}
\arguments{
\item{path}{Path to check; typically this will be your working
directory.}
}
\value{
Invisibly, a logical; \code{TRUE} if all checks succeed and
\code{FALSE} otherwise.
}
\description{
Perform some basic checks to make that your system is configured
to use the DIDE cluster properly.  Calling this when something
goes wrong is never a bad idea.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

dide_check()
\dontshow{\}) # examplesIf}
}
