% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dide.R
\name{dide_username}
\alias{dide_username}
\title{Report DIDE username}
\usage{
dide_username()
}
\value{
Your username, as a string
}
\description{
Report the username used to log into the web portal for use with
the DIDE cluster.  This may or may not be the same as your
local username.  We may ask you to run this when helping debug
cluster failures.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Return your DIDE username
dide_username()
\dontshow{\}) # examplesIf}
}
