% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{envvars}
\alias{envvars}
\title{Environment variables}
\usage{
envvars(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named environment
variable.  If unnamed, it is assumed to refer to an environment
variable that exists.}
}
\value{
A list with class \code{envvars} which should not be modified.
}
\description{
Create environment variables for use with a hipercow task.
}
