% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{hermod_configure}
\alias{hermod_configure}
\title{Configure your hermod root}
\usage{
hermod_configure(driver, ..., root = NULL)
}
\arguments{
\item{driver}{The hermod driver; probably you want this to be
\code{"windows"} as that is all we support at the moment!}

\item{...}{Arguments passed to your driver. We'll work out how to
point you at appropriate documentation once it is written.}

\item{root}{Hermod root, usually best \code{NULL}}
}
\description{
Configure your hermod root.  \code{hermod_configure} creates the
configuration and \code{hermod_get_configuration} looks it up.
}
