% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{hermod_init}
\alias{hermod_init}
\title{Create a hermod root}
\usage{
hermod_init(root = ".", driver = NULL, ...)
}
\arguments{
\item{root}{The path to the root, defaulting the current
directory.}

\item{driver}{Optionally, the name of a driver to configure}

\item{...}{Arguments passed through to \link{hermod_configure} if
\code{driver} is non-NULL.}
}
\value{
Invisibly, the root object
}
\description{
Create a hermod root.  This marks the directory where your task
information will be saved, along with a local copy of your R
packages (a "library" for the cluster).  Immediately after running
this the first time, you probably want to run
\code{\link[=hermod_configure]{hermod_configure()}} in order to control how we set up
your projects network paths and R version.
}
