% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{hermod_task_cancel}
\alias{hermod_task_cancel}
\title{Cancel tasks}
\usage{
hermod_task_cancel(id, root = NULL)
}
\arguments{
\item{id}{The task id or task ids to cancel}

\item{root}{A hermod root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A logical vector the same length as \code{id} indicating if the
task was cancelled. This will be \code{FALSE} if the job was already
completed, not running, etc.
}
\description{
Cancel one or more tasks
}
