% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{hermod_task_create_explicit}
\alias{hermod_task_create_explicit}
\title{Create explicit task}
\usage{
hermod_task_create_explicit(
  expr,
  export = NULL,
  envir = .GlobalEnv,
  packages = NULL,
  root = NULL
)
}
\arguments{
\item{expr}{Unevaluated expression object, e.g., from \code{quote}}

\item{export}{Optional character vector of names of objects to
export into the evaluating environment}

\item{envir}{Environment in which to find variables for \code{export}}

\item{packages}{Optional character vector of packages to attach
before the expression is run. Likely needed for things like
ggplot and dplyr which make use of exported functions.}

\item{root}{A hermod root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A task id, a string of hex characters. Use this to
interact with the task.
}
\description{
Create an explicit task. Explicit tasks are the simplest sort of
task in hermod and do nothing magic. They accept an R expression
(from \code{quote} or friends) and possibly a set of variables to
export from the global environment.  This can then be run on a
cluster by loading your variables and running your expression.  If
your expression depends on packages being \emph{attached} then you
should pass a vector of package names too.
}
