% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{hermod_task_create_expression}
\alias{hermod_task_create_expression}
\title{Create a task based on an expression}
\usage{
hermod_task_create_expression(expr, environment = "default", root = NULL)
}
\arguments{
\item{expr}{The expression, does not need quoting.}

\item{environment}{Name of the hermod environment to evaluate the
task within.}

\item{root}{A hermod root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A task id, a string of hex characters. Use this to
interact with the task.
}
\description{
Create a task based on an expression. This is similar to
\link{hermod_task_create_explicit} except more magic, and is closer to
the interface that we expect people will use.
}
