% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{hermod_task_driver}
\alias{hermod_task_driver}
\title{Return driver used by a task}
\usage{
hermod_task_driver(id, root = NULL)
}
\arguments{
\item{id}{The task identifier}

\item{root}{A hermod root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A string, \code{NA} if no driver used.
}
\description{
Return the driver used by a task. This can't be changed once set.
}
