% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{hermod_task_eval}
\alias{hermod_task_eval}
\title{Run a task}
\usage{
hermod_task_eval(id, envir = .GlobalEnv, root = NULL)
}
\arguments{
\item{id}{The task identifier}

\item{envir}{An environment in which to evaluate the
expression. For non-testing purposes, generally ignore this, the
global environment will be likely the expected environment.}

\item{root}{A hermod root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
Boolean indicating success (\code{TRUE}) or failure (\code{FALSE})
}
\description{
Run a task that has been created by a \verb{hermod_task_create_*}
function, e.g., \code{\link[=hermod_task_create_explicit]{hermod_task_create_explicit()}}
}
