% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{hermod_task_result}
\alias{hermod_task_result}
\title{Get task result}
\usage{
hermod_task_result(id, root = NULL)
}
\arguments{
\item{id}{The task identifier}

\item{root}{A hermod root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
The value of the queued expression
}
\description{
Get the task result. This might be an error if the task has failed.
}
