% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{hermod_task_status}
\alias{hermod_task_status}
\title{Get task status}
\usage{
hermod_task_status(id, root = NULL)
}
\arguments{
\item{id}{The task identifier}

\item{root}{A hermod root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A string with the task status. Tasks that do not exist
will have a status of \code{NA}.
}
\description{
Get the status of a task. See Details for the lifecycle.
}
\details{
A task passes through a lifecycle:
\itemize{
\item \code{created}
\item \code{submitted}
\item \code{started}
\item \code{success}, \code{failure}, \code{cancelled}
}

These occur in increasing order and the result of this function is
the furtherst through this list.

Later, we will introduce other types to cope with tasks that have
been retried, or that are blocked on dependencies (or have become
impossible).
}
