% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{hermod_task_submit}
\alias{hermod_task_submit}
\title{Submit a task}
\usage{
hermod_task_submit(id, ..., driver = NULL, root = NULL)
}
\arguments{
\item{id}{The task id}

\item{...}{Disallowed additional arguments, don't use.}

\item{driver}{The name of the driver to use, or you can leave
blank if only one is configured (this will be typical).}

\item{root}{The hermod root}
}
\description{
Submit a task to a queue
}
\details{
This is a lower-level function that you will not often need to call.
}
