% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_cancel}
\alias{hipercow_bundle_cancel}
\title{Cancel bundle tasks}
\usage{
hipercow_bundle_cancel(bundle, follow = TRUE, root = NULL)
}
\arguments{
\item{bundle}{Either a \code{hipercow_bundle} object, or the name of a
bundle.}

\item{follow}{Logical, indicating if we should follow any retried
tasks.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A logical vector the same length as \code{id} indicating if the
task was cancelled. This will be \code{FALSE} if the job was already
completed, not running, etc.
}
\description{
Cancel all tasks in a bundle.  This wraps \link{task_cancel} for all
the ids.
}
\examples{
cleanup <- hipercow_example_helper(runner = FALSE)

bundle <- task_create_bulk_expr(sqrt(x), data.frame(x = 1:5))
hipercow_bundle_cancel(bundle)
hipercow_bundle_status(bundle)

cleanup()
}
