% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_create}
\alias{hipercow_bundle_create}
\title{Create task bundle}
\usage{
hipercow_bundle_create(
  ids,
  name = NULL,
  validate = TRUE,
  overwrite = TRUE,
  root = NULL
)
}
\arguments{
\item{ids}{A character vector of task ids}

\item{name}{A string, the name for the bundle.  If not given, then
a random nmae is generated.}

\item{validate}{Logical, indicating if we should check that the
task ids exist.  We always check that the task ids are
plausible.}

\item{overwrite}{Logical, indicating that we should overwrite any
existing bundle with the same name.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A task bundle object
}
\description{
Create a bundle of tasks.  This is simply a collection of tasks
that relate together in some way, and we provide some helper
functions for working with them that save you writing lots of
loops.  Each bundle has a name, which will be randomly generated
if you don't provide one, and a set of task ids.
}
