% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_delete}
\alias{hipercow_bundle_delete}
\title{Delete task bundles}
\usage{
hipercow_bundle_delete(name, root = NULL)
}
\arguments{
\item{name}{Character vectors of names to delete}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
Nothing, called for its side effect
}
\description{
Delete one or more hipercow task bundles.  Note that this does not
delete the underlying tasks, which is not yet supported.
}
\examples{
hipercow_example_helper()

bundle <- task_create_bulk_expr(sqrt(x), data.frame(x = 1:5))
hipercow_bundle_list()

# Retaining the ids, delete bundle
ids <- bundle$ids
hipercow_bundle_delete(bundle$name)
hipercow_bundle_list()

# The tasks still exist:
task_status(ids)
}
