% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_list}
\alias{hipercow_bundle_list}
\title{List existing bundles}
\usage{
hipercow_bundle_list(root = NULL)
}
\arguments{
\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A \link{data.frame} with columns \code{name} and \code{time}, ordered by
time (most recent first)
}
\description{
List existing bundles
}
