% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_load}
\alias{hipercow_bundle_load}
\title{Load existing bundle}
\usage{
hipercow_bundle_load(name, root = NULL)
}
\arguments{
\item{name}{Name of the bundle to load}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A \code{hipercow_bundle} object
}
\description{
Load an existing saved bundle
}
