% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_log_value}
\alias{hipercow_bundle_log_value}
\title{Fetch bundle logs}
\usage{
hipercow_bundle_log_value(bundle, outer = FALSE, follow = TRUE, root = NULL)
}
\arguments{
\item{bundle}{Either a \code{hipercow_bundle} object, or the name of a
bundle.}

\item{outer}{Logical, indicating if we should request the "outer"
logs; these are logs from the underlying HPC software before it
hands off to hipercow.}

\item{follow}{Logical, indicating if we should follow any retried
tasks.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A list with each element being the logs for the
corresponding element in the bundle.
}
\description{
Fetch logs from tasks in a bundle.
}
\examples{
cleanup <- hipercow_example_helper(with_logging = TRUE)
bundle <- task_create_bulk_expr(sqrt(x), data.frame(x = 1:2))
hipercow_bundle_wait(bundle)
hipercow_bundle_log_value(bundle)

cleanup()
}
