% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_status}
\alias{hipercow_bundle_status}
\title{Bundle status}
\usage{
hipercow_bundle_status(bundle, reduce = FALSE, follow = TRUE, root = NULL)
}
\arguments{
\item{bundle}{Either a \code{hipercow_bundle} object, or the name of a
bundle.}

\item{reduce}{Reduce the status across all tasks in the bundle.
This means we return a single value with the "worst" status
across the bundle.  We only return \code{success} if \emph{all} tasks have
succeeded, and will return \code{failed} if any task has failed.}

\item{follow}{Logical, indicating if we should follow any retried
tasks.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A character vector the same length as the number of tasks
in the bundle, or length 1 if \code{reduce} is \code{TRUE}.
}
\description{
Fetch status for all tasks in a bundle.
}
\examples{
hipercow_example_helper()
bundle <- task_create_bulk_expr(sqrt(x), data.frame(x = 1:5))
# Immediately after submission, tasks may not all be complete so
# we may get a mix of statuses.  In that case the reduced status
# will be "submitted" or "running", even though some tasks may be
# "success"
hipercow_bundle_status(bundle)
hipercow_bundle_status(bundle, reduce = TRUE)

# After completion all tasks have status "success", as does the
# reduction.
hipercow_bundle_wait(bundle)
hipercow_bundle_status(bundle)
hipercow_bundle_status(bundle, reduce = TRUE)
}
