% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{hipercow_cluster_info}
\alias{hipercow_cluster_info}
\title{Retrieve cluster information}
\usage{
hipercow_cluster_info(driver = NULL, root = NULL)
}
\arguments{
\item{driver}{The name of the driver to use, or you can leave
blank if only one is configured (this will be typical).}

\item{root}{The hipercow root}
}
\value{
A list of properties
}
\description{
Retrieve info from a driver about its cluster's capabilities; the
maximum number of cores, maximum memory, queues and node lists.
}
