% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{hipercow_configuration}
\alias{hipercow_configuration}
\title{Report on hipercow configuration}
\usage{
hipercow_configuration(show = TRUE, root = NULL)
}
\arguments{
\item{show}{Display the configuration to the screen}

\item{root}{Hipercow root, usually best \code{NULL}}
}
\value{
A list with a machine readable form of this information,
invisibly.
}
\description{
Report on your hipercow configuration.  We will always want you to
post this along side any problems; it has lots of useful
information in it that will help us see how your set up is
configured.
}
