% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{hipercow_configure}
\alias{hipercow_configure}
\title{Configure your hipercow root}
\usage{
hipercow_configure(driver, ..., root = NULL)
}
\arguments{
\item{driver}{The hipercow driver; probably you want this to be
\code{"windows"} as that is all we support at the moment!}

\item{...}{Arguments passed to your driver. We'll work out how to
point you at appropriate documentation once it is written.}

\item{root}{Hipercow root, usually best \code{NULL}}
}
\description{
Configure your hipercow root.  \code{hipercow_configure} creates the
configuration and \code{hipercow_configuration} looks it up.
}
