% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hipercow_hello}
\alias{hipercow_hello}
\title{Hello world}
\usage{
hipercow_hello(progress = NULL, timeout = Inf, root = NULL)
}
\arguments{
\item{progress}{Logical value, indicating if a progress spinner
should be used. The default \code{NULL} uses the option
\code{hipercow.progress}, and if unset displays a progress bar in an
interactive session.}

\item{timeout}{The time to wait for the task to complete. The
default is to wait forever.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
Unclear
}
\description{
Hello world in hipercow. This function sends a tiny test task
through the whole system to confirm that everything is configured
correctly.
}
