% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{hipercow_parallel}
\alias{hipercow_parallel}
\title{Specify parallel use of cores}
\usage{
hipercow_parallel(method = NULL)
}
\arguments{
\item{method}{The parallel method that hipercow will prepare.
Three options are available: the \code{future} package, the \code{parallel}
page, or \code{NULL}, the default, will do nothing. See the details for
examples.}
}
\value{
A list containing your parallel configuration.
}
\description{
Set parallel options.  Having requested more than one core using
\link{hipercow_resources}, here hipercow can start up a local cluster
on the node you are running on, using either the \code{future} or
\code{parallel} package.
}
\details{
For example, using \code{future}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{resources <- hipercow_resources(cores = 4)
id <- task_create_expr(
  furrr::future_map(1:4, ~Sys.getpid()), 
  parallel = hipercow_parallel("future"),
  resources = resources)
}\if{html}{\out{</div>}}

where \code{furrr} must be provisioned using \link{hipercow_provision}.

For use with \code{parallel}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{id <- task_create_expr(
  parallel::clusterApply(NULL, 1:4, function(x) Sys.getpid()), 
  parallel = hipercow_parallel("parallel"),
  resources = resources)
}\if{html}{\out{</div>}}
}
