% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{hipercow_parallel_get_cores}
\alias{hipercow_parallel_get_cores}
\title{Get number of cores}
\usage{
hipercow_parallel_get_cores()
}
\value{
The number of cores a cluster has allocated to your task. This will
be less than or equal to the number of cores on the cluster node
running your task.
}
\description{
Lookup number of cores allocated to the task
}
