% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{hipercow_parallel_load_environment}
\alias{hipercow_parallel_load_environment}
\title{Load an environment in a parallel context}
\usage{
hipercow_parallel_load_environment(name, envir = .GlobalEnv)
}
\arguments{
\item{envir}{The name of the environment}
}
\value{
Nothing, called for side effects only
}
\description{
Load an environment into a parallel worker.  This is a helper
function designed for use from parallel backends, and is exported
mostly because it makes it easier for us to work with.  Users
should never need to call this function directly.
}
\keyword{internal}
