% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{hipercow_resource}
\alias{hipercow_resource}
\title{Hipercow Resources}
\usage{
hipercow_resource(
  cores = 1,
  exclusive = FALSE,
  runtime = NULL,
  hold_until = NULL,
  memory_per_node = NULL,
  memory_per_process = NULL,
  requested_nodes = NULL,
  priority = NULL,
  queue = NULL
)
}
\arguments{
\item{cores}{The number of cores your task requires. This is 1 by
default. An error is reported if you request more cores than
is available on the largest cluster node. Setting to \code{Inf} will
request any entire node, however many cores it has.}

\item{exclusive}{Set this to \code{TRUE} to ensure no other tasks will be
concurrently run on the node while it runs your task.}

\item{runtime}{Set this to specify a time limit for running your job.
Acceptable formats are either an integer number of minutes, or
characters in the form \code{"m"}, \code{"h:m"}, or \code{"d:h:m"} to optionally
specify days, hours and minutes.}

\item{hold_until}{POSIXt, Date, the special strings "tonight",
"midnight", "weekend" (midnight Saturday morning).

Specify your task should wait in the queue,
rather than start running as soon as possible. Options are, an
integer number of minutes, characters in the form \code{"m"}, or \code{"h:m"}
for hours and minutes, or \code{"tonight"} for 7pm, or \code{"midnight"} for
midnight.}

\item{memory_per_node}{Specify your task can only run on a node
with at least the specified memory. This is an integer,
measured in gigabytes, or strings such as \code{"64Gb"} or \code{"1Tb"}.
Our latest nodes have either 384Gb, 512Gb, or 1Tb of RAM.}

\item{memory_per_process}{If you can provide an estimate of how
much RAM your task requires, then the cluster can ensure the
total memory required by running multiple tasks on a node
does not exceed how much memory the node has. Specify this as
an integer number of gigabytes, or characters such as \code{"10Gb"}}

\item{requested_nodes}{If you have been in touch with us or DIDE
IT, and you need to run your task on a selection of named compute
nodes, then specify this here as a vector of node names.}

\item{priority}{If the tasks you are launching can go on the
back-burner and have other tasks jump over them, then specify
priority to \code{low}; otherwise \code{normal}. These are the only
acceptable values.}

\item{queue}{Specify a particular queue to have your tasks submit
to. This is in development as we'll decide in the future what
queues we best need for DIDE's common workflows. See the Details
for more information}

\item{headnode}{At present, we are only supporting the new \code{wpia-hn}
headnode, but hipercow may support others in the future.}
}
\value{
A hipercow_resource list, which can be passed to
hypercow_configure
}
\description{
Specify what resources a job requires to run.
}
\section{Windows cluster}{
The windows cluster currently is all 32-core nodes, with RAM
mainly 512Gb, with a few 384Gb.
}

\section{Linux cluster}{
Coming Soon.
}

