% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{hipercow_resources_validate}
\alias{hipercow_resources_validate}
\title{Validate a \code{hipercow_resource} for a driver.}
\usage{
hipercow_resources_validate(resources, driver = NULL, root = NULL)
}
\arguments{
\item{resources}{A \code{hipercow_resource} list returned by
\code{hipercow_resources}.}

\item{driver}{The name of the driver to use, or you can leave
blank if only one is configured (this will be typical).}

\item{root}{The hipercow root}
}
\value{
TRUE if the resources are compatible with this driver.
}
\description{
Query a driver to find information about the cluster, and then validate
a \code{hipercow_resource} list against that driver to see if the resources
requested could be satisfied.
}
