% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq.R
\name{hipercow_rrq_stop_workers_once_idle}
\alias{hipercow_rrq_stop_workers_once_idle}
\title{Tell workers to exit once complete}
\usage{
hipercow_rrq_stop_workers_once_idle(root = NULL)
}
\arguments{
\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
Nothing, called for side effects only
}
\description{
Tell workers to exit once work is complete
}
