% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{hipercow_unconfigure}
\alias{hipercow_unconfigure}
\title{Remove a driver from a hipercow configuration}
\usage{
hipercow_unconfigure(driver, root = NULL)
}
\arguments{
\item{driver}{The name of the driver to remove}

\item{root}{Hipercow root, usually best \code{NULL}}
}
\value{
Nothing, called for its side effects only.
}
\description{
Remove a driver configured by \link{hipercow_configure}.  This will not
affect tasks already submitted with this driver, but will prevent
any future tasks being submitted with it.
}
