% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_create_explicit}
\alias{task_create_explicit}
\title{Create explicit task}
\usage{
task_create_explicit(
  expr,
  export = NULL,
  envir = .GlobalEnv,
  environment = "default",
  root = NULL
)
}
\arguments{
\item{expr}{Unevaluated expression object, e.g., from \code{quote}}

\item{export}{Optional character vector of names of objects to
export into the evaluating environment}

\item{envir}{Local R environment in which to find variables for
\code{export}}

\item{environment}{Name of the hermod environment to evaluate the
task within.}

\item{root}{A hermod root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A task id, a string of hex characters. Use this to
interact with the task.
}
\description{
Create an explicit task. Explicit tasks are the simplest sort of
task in hermod and do nothing magic. They accept an R expression
(from \code{quote} or friends) and possibly a set of variables to
export from the global environment.  This can then be run on a
cluster by loading your variables and running your expression.  If
your expression depends on packages being \emph{attached} then you
should pass a vector of package names too.  This function may
disappear, and is used by us to think about the package, it's not
designed to really be used.
}
