% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-list.R
\name{task_list}
\alias{task_list}
\title{List tasks}
\usage{
task_list(
  task_ids = NULL,
  finished_before = NULL,
  in_bundle = NULL,
  with_status = NULL,
  root = NULL
)
}
\arguments{
\item{task_ids}{A character vector of task identifiers.  Typically
if you provide this you will not provide any other filters.}

\item{finished_before}{A date, time, or \link{difftime} object
representing the time or time ago that a task finished (here,
the job might have finished for any reason; successfully or
unsuccessfully unless you also provide the \code{with_status}
argument).}

\item{in_bundle}{A character vector of bundle names. Wild cards
are supported using shell (glob) syntax, rather than regular
expression syntax.  So use \verb{data_*} to match all bundles that
start with \code{data_} (see \link[utils:glob2rx]{utils::glob2rx} for details).  It is an
error if \emph{no} bundles are matched, but not an error if any
individual pattern does not match.}

\item{with_status}{A character vector of statuses to match.  We
only purge tasks that match these statuses.  Valid statuses to
use are \code{created}, \verb{submitted, }running\verb{, }success\verb{, }failure\code{and}cancelled`.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A character vector.  You may want to then pull this vector
of ids into a bundle (e.g., \link{hipercow_bundle_create}).  The
order is arbitrary and does not reflect anything in your tasks.
}
\description{
List hipercow tasks.  This is rarely what you want to, but can be
useful in some situations.  We get requests for this function
periodically, so here it is!  Other, better, functions may be
available (see Details).  Please be aware that this function can
be quite slow, and do not overuse it in scripts (please do not,
for example, call this repeatedly in a loop - talk to us if you
are tempted to do this).
}
\details{
Sometimes, better functions are available for you:
\itemize{
\item If you want to list tasks in order to delete them, you might
prefer \code{\link[=hipercow_purge]{hipercow_purge()}}
\item If you want to list tasks in a bundle, you should use
\code{\link[=hipercow_bundle_list]{hipercow_bundle_list()}} to find the bundle and
\code{\link[=hipercow_bundle_load]{hipercow_bundle_load()}} to load it (or use other bundle
functions)
}

Once you have listed tasks with \code{task_list()}
}
