% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{task_submit}
\alias{task_submit}
\title{Submit a task}
\usage{
task_submit(id, ..., driver = NULL, root = NULL)
}
\arguments{
\item{id}{A vector of task ids}

\item{...}{Disallowed additional arguments, don't use.}

\item{driver}{The name of the driver to use, or you can leave
blank if only one is configured (this will be typical).}

\item{root}{The hipercow root}
}
\description{
Submit a task to a queue.  This is a lower-level function that you
will not often need to call.  Typically a task will be submitted
automatically to your driver on creation (e.g., with
\code{\link[=task_create_expr]{task_create_expr()}}), unless you specified \code{submit = FALSE} or you had not yet configured a driver.
}
