% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_wait}
\alias{task_wait}
\title{Wait for a task to complete}
\usage{
task_wait(id, timeout = Inf, poll = 1, progress = NULL, root = NULL)
}
\arguments{
\item{id}{The task identifier}

\item{timeout}{The time to wait for the task to complete. The
default is to wait forever.}

\item{poll}{The interval to request an update; shorter values here
will return more quickly but will flail around a bit more.  The
default is to check every second which means you wait for up to
a second longer than needed - for long running jobs you could
set this longer if you wanted.}

\item{progress}{Logical value, indicating if a progress indicator
should be used. The default \code{NULL} uses the option
\code{hipercow.progress}, and if unset displays a progress bar in an
interactive session.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
Logical value, \code{TRUE} if the task completed successfully,
\code{FALSE} otherwise.
}
\description{
Wait for a single task to complete.
}
