% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{windows_authenticate}
\alias{windows_authenticate}
\title{DIDE windows credentials}
\usage{
windows_authenticate()
}
\value{
Nothing, this function is called for its side effect of
setting or updating your credentials within the keyring.
}
\description{
Register DIDE windows credentials.
}
\details{
In order to be able to communicate with the Windows DIDE HPC
system, we need to be able to communicate with the HPC portal
(\url{https::mrcdata.dide.ic.ac.uk/hpc}), and for this we need your
\strong{DIDE} password and username. This is typically, but not always,
the same as your Imperial credentials.  We store this information
securely using the \href{https://keyring.r-lib.org/}{keyring} package,
so when unlocking your credentials you will be prompted for your
\strong{computer} password, which will be your DIDE password if you use
a windows machine connected to the DIDE domain, but will likely
differ from either your DIDE or Imperial password if you are
outside the DIDE domain, or if you don't use Windows.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

windows_authenticate()
\dontshow{\}) # examplesIf}
}
