% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{windows_check}
\alias{windows_check}
\title{Check we can use windows cluster}
\usage{
windows_check()
}
\value{
Invisibly, a boolean; \code{TRUE} if all checks succeed and
\code{FALSE} otherwise.
}
\description{
Perform some basic checks to make that your system is configured
to use the windows cluster properly.
}
