% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{windows_generate_keypair}
\alias{windows_generate_keypair}
\title{Generate keypair}
\usage{
windows_generate_keypair(update = FALSE)
}
\arguments{
\item{update}{Replace the existing keypair.  You will need to use
this if you accidentally remove the \verb{.hipercow/} directory from
your network home share, or if you want to renew your key.}
}
\value{
Nothing, called for its side effect
}
\description{
Generate a keypair for encrypting small data to send to the
windows cluster.  This can be used to encrypt environment
variables, and possibly other workflows in future.  By default, if
you have ever created a keypair we do not replace it if it already
exists, unless you set \code{update = TRUE} so you may call this
function safely to ensure that you do have a keypair set up.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Generate a new keypair, if one does not exist
windows_generate_keypair()
\dontshow{\}) # examplesIf}
}
