% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{windows_path}
\alias{windows_path}
\title{Describe a path mapping}
\usage{
windows_path(path_local, path_remote, drive_remote, call = NULL)
}
\arguments{
\item{path_local}{The point where the drive is attached locally.
On Windows this will be something like "Q:/", on Mac something
like "/Volumes/mountname", and on Linux it could be anything at
all, depending on what you used when you mounted it (or what is
written in \verb{/etc/fstab})}

\item{path_remote}{The network path for this drive.  It
will look something like \verb{\\\\\\\\projects.dide.ic.ac.uk\\\\tmp\\\\}.
Unfortunately backslashes are really hard to get right here and
you will need to use twice as many as you expect (so \emph{four}
backslashes at the beginning and then two for each separator).
If this makes you feel bad know that you are not alone:
https://xkcd.com/1638 -- alternatively you may use forward
slashes in place of backslashes (e.g. \verb{//projects.dide.ic.ac.uk/tmp})}

\item{drive_remote}{The place to mount the drive on the cluster.
We're probably going to mount things at Q: and T: already so
don't use those.  And things like C: are likely to be used.
Perhaps there are some guidelines for this somewhere?}

\item{call}{The name of the calling function, for error reporting.}
}
\description{
Describe a path mapping for use when setting up jobs on the cluster.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Suppose you have mounted your malaria share at "~/net/malaria"
# (e.g., on a Linux machine).  You can tell the cluster to mount
# this as "M:" when running tasks by first creating a path
# mapping:
share <- windows_path("~/net/malaria",
                      "//wpia-hn.hpc.dide.ic.ac.uk/Malaria",
                      "M:")

# This share object contains information about how to relate your
# local and remote paths:
share

# When configuring the cluster you might pass this:
hipercow_configure("windows", shares = share)
\dontshow{\}) # examplesIf}
}
