% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{windows_username}
\alias{windows_username}
\title{Report windows username}
\usage{
windows_username()
}
\value{
Your username, as a string
}
\description{
Report the username used to log into the web portal for use with
the windows cluster.  This may or may not be the same as your
local username.  We may ask you to run this when helping debug
cluster failures.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Return your windows username
windows_username()
\dontshow{\}) # examplesIf}
}
