% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{hermod_configure}
\alias{hermod_configure}
\title{Configure your hermod root}
\usage{
hermod_configure(shares = NULL, r_version = NULL, root = NULL)
}
\arguments{
\item{shares}{Network shares in addition to your working
directory. You only need to add this if you want to access some
files from a network location other than the one that you are
working in; e.g., you have large files that you want to read on
your \verb{P:}, you could map that here, and then refer to
\verb{P:/myfiles/} in your job.}

\item{r_version}{A string, or \code{numeric_version} object, describing
the R version required.  Not all R versions are known to be
supported, so this will check against a list of installed R
versions for the cluster you are using.  If omitted then: if
your R version matches a version on the cluster that will be
used, or the oldest cluster version that is newer than yours, or
the most recent cluster version.}

\item{root}{Hermod root, usually best \code{NULL}}
}
\description{
Configure your hermod root.
}
\author{
Richard Fitzjohn
}
