% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{hermod_driver}
\alias{hermod_driver}
\title{Create a new hermod driver; this is intended to be used from other
packages, and rarely called directly. If you are trying to run
tasks on a cluster you do not need to call this!}
\usage{
hermod_driver(configure, submit, status)
}
\arguments{
\item{configure}{Function used to set core configuration for the
driver.  This function will be called from the hermod root
directory (so \code{getwd()} will report the correct path). It can
take any arguments, do any calculation and then must return any
R object that can be serialised.  The resulting configuration
will be passed in as \code{config} to other driver functions.}

\item{submit}{Submit a task to a cluster.  This is run after the
task is created (either automatically or manually) and takes as
arguments the task id, the configuration, the path to the root.}

\item{status}{Fetch a task status. Takes a vector of ids and
returns a vector of the same length of statuses.}
}
\description{
Create a new hermod driver; this is intended to be used from other
packages, and rarely called directly. If you are trying to run
tasks on a cluster you do not need to call this!
}
