% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provision.R
\name{hermod_provision}
\alias{hermod_provision}
\title{Provision cluster library}
\usage{
hermod_provision(method = NULL, ..., driver = NULL, root = NULL)
}
\arguments{
\item{method}{The provisioning method to use, defaulting to
\code{NULL}, which indicates autoprovision}

\item{...}{Arguments passed through to conan. See docs that we
need to write still.}

\item{driver}{The name of the driver to use, or you can leave
blank if only one is configured (this will be typical).}

\item{root}{The hermod root}
}
\value{
Nothing
}
\description{
Provision a library. Normally we will do this automatically for
you, but with this function you can trigger provisioning.
}
