% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_delete}
\alias{hipercow_bundle_delete}
\title{Delete task bundles}
\usage{
hipercow_bundle_delete(name, root = NULL)
}
\arguments{
\item{name}{Character vectors of names to delete}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
Nothing, called for its side effect
}
\description{
Delete one or more hipercow task bundles
}
