% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_list}
\alias{hipercow_bundle_list}
\title{List existing bundles}
\usage{
hipercow_bundle_list(root = NULL)
}
\arguments{
\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A \link{data.frame} with columns \code{name} and \code{time}, ordered by
time (most recent first)
}
\description{
List existing bundles
}
\examples{
hipercow_example_helper()

# With no bundles present
hipercow_bundle_list()

# With a bundle
bundle <- task_create_bulk_expr(sqrt(x), data.frame(x = 1:5))
hipercow_bundle_list()
}
