% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_load}
\alias{hipercow_bundle_load}
\title{Load existing bundle}
\usage{
hipercow_bundle_load(name, root = NULL)
}
\arguments{
\item{name}{Name of the bundle to load}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A \code{hipercow_bundle} object
}
\description{
Load an existing saved bundle by name.  This is intended for where
you have created a long-running bundle and since closed down your
session.  See \link{hipercow_bundle_list} for finding names of bundles.
}
\examples{
cleanup <- hipercow_example_helper()

bundle <- task_create_bulk_expr(sqrt(x), data.frame(x = 1:5))
name <- bundle$name

# Delete the bundle object; the bundle exists still in hipercow's store.
rm(bundle)

# With the name we can load the bundle and fetch its status
bundle <- hipercow_bundle_load(name)
hipercow_bundle_status(bundle)

# In fact, you can use just the name if you prefer:
hipercow_bundle_status(name)

cleanup()
}
