% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_log_value}
\alias{hipercow_bundle_log_value}
\title{Fetch bundle logs}
\usage{
hipercow_bundle_log_value(bundle, follow = TRUE, outer = FALSE, root = NULL)
}
\arguments{
\item{bundle}{Either a \link{hipercow_bundle} object, or the name of a
bundle.}

\item{follow}{Logical, indicating if we should follow any retried
tasks.}

\item{outer}{Logical, indicating if we should request the "outer"
logs; these are logs from the underlying HPC software before it
hands off to hipercow.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A list with each element being the logs for the
corresponding element in the bundle.
}
\description{
Fetch logs from tasks in a bundle.
}
