% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{hipercow_bundle_status}
\alias{hipercow_bundle_status}
\title{Bundle status}
\usage{
hipercow_bundle_status(bundle, reduce = FALSE, root = NULL)
}
\arguments{
\item{bundle}{Either a \link{hipercow_bundle} object, or the name of a
bundle.}

\item{reduce}{Reduce the status across all tasks in the bundle.
This means we return a single value with the "worst" status
across the bundle.  We only return \code{success} if \emph{all} tasks have
succeeded, and will return \code{failed} if any task has failed.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A character vector the same length as the number of tasks
in the bundle, or length 1 if \code{reduce} is \code{TRUE}.
}
\description{
Fetch status for all tasks in a bundle.
}
