% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{hipercow_cluster_info}
\alias{hipercow_cluster_info}
\title{Describe cluster}
\usage{
hipercow_cluster_info(driver = NULL, root = NULL)
}
\arguments{
\item{driver}{The driver to use, which determines the cluster to
fetch information from (depending on your configuration).  If no
driver is configured, an error will be thrown.}

\item{root}{Hipercow root, usually best \code{NULL}}
}
\value{
A list describing the cluster.  The details depend on the
driver, and are subject to change.  We expect to see elements:
\itemize{
\item resources: Describes the computational resources on the cluster,
which is used by \link{hipercow_resources_validate}.  Currently this
is a simple list with elements \code{max_ram} (max RAM available, in
GB), \code{max_cores} (max number of cores you can request), \code{queues}
(character vector of available queues), \code{nodes} (character
vector of available nodes), \code{default_queue} (the default queue).
These details are subject to change but the contents should
always be informative and fairly self explanatory.
\item redis_url: The URL of the redis server to communicate with from
outside of the cluster (i.e., from your computer), in a form
suitable for use with redux::hiredis
\item r_versions: A vector of R versions, as \code{numeric_vector} objects
}
}
\description{
Describe information about the cluster.  This is (naturally) very
dependent on the cluster but some details of the value are
reliable; see Value for details.
}
\examples{
cleanup <- hipercow_example_helper()
hipercow_cluster_info()
cleanup()
}
