% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvars.R
\name{hipercow_envvars}
\alias{hipercow_envvars}
\title{Environment variables}
\usage{
hipercow_envvars(..., secret = FALSE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named environment
variable.  If unnamed, it is assumed to refer to an environment
variable that exists.}

\item{secret}{Are these environment variables secret?  If so we will save}
}
\value{
A list with class \code{hipercow_envvars} which should not be modified.
}
\description{
Create environment variables for use with a hipercow task.
}
