% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{hipercow_example_helper}
\alias{hipercow_example_helper}
\title{Example helper}
\usage{
hipercow_example_helper(runner = TRUE, with_logging = FALSE, envir = NULL)
}
\arguments{
\item{runner}{Start a runner?}

\item{with_logging}{Run each task with logging; this is quite a
bit slower.}

\item{envir}{Environment to base lifetime on; the runner will be
stopped and working directory restored once this environment
exits.}
}
\description{
A helper used in running examples
}
\keyword{internal}
