% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{hipercow_example_helper}
\alias{hipercow_example_helper}
\title{Example helper}
\usage{
hipercow_example_helper(runner = TRUE, with_logging = FALSE)
}
\arguments{
\item{runner}{Start a runner?}

\item{with_logging}{Run each task with logging; this is quite a
bit slower.}
}
\description{
A helper used in running examples
}
\keyword{internal}
