% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{hipercow_example_helper}
\alias{hipercow_example_helper}
\title{Example helper}
\usage{
hipercow_example_helper(runner = TRUE, with_logging = FALSE)
}
\arguments{
\item{runner}{Start a runner?}

\item{with_logging}{Run each task with logging; this is quite a
bit slower.}
}
\value{
A function that can be called (with no arguments) to
return to the original working directory and clean up all files
created for the example.
}
\description{
A helper used in running examples and docs.  This function will
change your working directory into a new temporary hipercow root,
and start a worker process that will quietly run tasks.  This is
not intended for people to use outside of running examples!
}
\keyword{internal}
