% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hipercow_hello}
\alias{hipercow_hello}
\title{Hello world}
\usage{
hipercow_hello(progress = NULL, timeout = NULL, driver = NULL)
}
\arguments{
\item{progress}{Logical value, indicating if a progress spinner
should be used. The default \code{NULL} uses the option
\code{hipercow.progress}, and if unset displays a progress bar in an
interactive session.}

\item{timeout}{The time to wait for the task to complete. The
default is to wait forever.}

\item{driver}{The driver to use to send the test task.  This can
be omitted where you have exactly one driver, but we error if
not given when you have more than one driver, or if you have not
configured any drivers.}
}
\value{
The string "Moo", direct from your cluster.
}
\description{
Hello world in hipercow. This function sends a tiny test task
through the whole system to confirm that everything is configured
correctly.
}
\examples{

cleanup <- hipercow_example_helper()
hipercow_hello()

cleanup()
}
