% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{hipercow_parallel}
\alias{hipercow_parallel}
\title{Specify parallel use of cores}
\usage{
hipercow_parallel(method = NULL, ..., driver = NULL, root = NULL)
}
\arguments{
\item{method}{The parallel method that hipercow will prepare.
Four options are available: \code{future} will allocate your
reserved cores to a (furrr)\link{https://furrr.futureverse.org/}
cluster; \code{parallel} will make you a cluster using the
(parallel)\link{https://stat.ethz.ch/R-manual/R-devel/library/parallel/doc/parallel.pdf}
package; \code{doParallel} will initialise using the
(doParallel)\link{https://cran.r-project.org/web/packages/doParallel/vignettes/gettingstartedParallel.pdf}
package, and finally \code{NULL}, the default, will do nothing,
and leave you to configure any parallelism yourself.}

\item{...}{Disallowed additional arguments, don't use.}

\item{driver}{The name of the driver to use, or you can leave
blank if only one is configured (this will be typical).}

\item{root}{The hipercow root}
}
\value{
A list containing your parallel configuration.
}
\description{
Set parallel options.  Having requested more than one core using
\link{hipercow_resources}, here we specify the way that those cores
might be used in parallel.
}
