% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{hipercow_parallel}
\alias{hipercow_parallel}
\title{Specify parallel use of cores}
\usage{
hipercow_parallel(method = NULL)
}
\arguments{
\item{method}{The parallel method that hipercow will prepare.
Four options are available: \code{future} will allocate your
reserved cores to a \href{https://furrr.futureverse.org/}{furrr}
cluster; \code{parallel} will make you a cluster using the
\href{https://www.rdocumentation.org/packages/parallel/}{parallel}
package; \code{doParallel} will initialise using the
\href{https://cran.r-project.org/web/packages/doParallel/index.html}{doParallel}
package, and finally \code{NULL}, the default, will do nothing,
and leave you to configure any parallelism yourself. The packages you
use should be provisioned in the usual way with \link{hypercow_provision}.}
}
\value{
A list containing your parallel configuration.
}
\description{
Set parallel options.  Having requested more than one core using
\link{hipercow_resources}, here we specify the way that those cores
might be used in parallel.
}
