% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{hipercow_resources_validate}
\alias{hipercow_resources_validate}
\title{Validate a \code{hipercow_resources} list for a driver.}
\usage{
hipercow_resources_validate(
  resources,
  platform = "windows",
  driver = NULL,
  root = NULL
)
}
\arguments{
\item{resources}{A \link{hipercow_resources} list returned by
\link{hipercow_resources}, or \code{NULL}}

\item{driver}{The name of the driver to use, or you can leave
blank if only one is configured (this will be typical).}

\item{root}{The hipercow root}
}
\value{
TRUE if the resources are compatible with this driver.
}
\description{
Query a driver to find information about the cluster, and then validate
a \link{hipercow_resources} list against that driver to see if the resources
requested could be satisfied.
}
\examples{
cleanup <- hipercow_example_helper()
hipercow_resources_validate(hipercow_resources(cores = 1))

# This example does not allow more than one core
tryCatch(
  hipercow_resources_validate(hipercow_resources(cores = 32)),
  error = identity)

cleanup()
}
