% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrq.R
\name{hipercow_rrq_controller}
\alias{hipercow_rrq_controller}
\title{Create an rrq controller}
\usage{
hipercow_rrq_controller(..., set_as_default = TRUE, driver = NULL, root = NULL)
}
\arguments{
\item{...}{Additional arguments passed through to
\code{\link[rrq:rrq_controller]{rrq::rrq_controller()}}; currently this is \code{follow} and
\code{timeout_task_wait}.}

\item{set_as_default}{Set the rrq controller to be the default;
this is usually what you want.}

\item{driver}{Name of the driver to use.  The default (\code{NULL})
depends on your configured drivers; if you have no drivers
configured we will error as we lack information required to
proceed.  If you have exactly one driver configured we'll submit
your task with it.  If you have more than one driver configured,
then we will error, though in future versions we may fall back
on a default driver if you have one configured.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
An \link[rrq:rrq_controller]{rrq::rrq_controller} object.
}
\description{
Create an rrq controller for your queue, and set it as the default
controller.  Use this to interact with workers created with
\code{\link[=hipercow_rrq_workers_submit]{hipercow_rrq_workers_submit()}}.  Proper docs forthcoming, all
interfaces are subject to some change.
}
