% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_cancel}
\alias{task_cancel}
\title{Cancel tasks}
\usage{
task_cancel(id, follow = TRUE, root = NULL)
}
\arguments{
\item{id}{The task id or task ids to cancel}

\item{follow}{Logical, indicating if we should follow any retried
tasks.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A logical vector the same length as \code{id} indicating if the
task was cancelled. This will be \code{FALSE} if the task was already
completed, not running, etc.
}
\description{
Cancel one or more tasks
}
\examples{
hipercow_example_helper()

ids <- c(task_create_expr(Sys.sleep(2)), task_create_expr(runif(1)))

# The first task may or not be cancelled (depends on if it was
# started already) but the second one will almos certainly be
# cancelled:
task_cancel(ids)
}
