% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_create_script}
\alias{task_create_script}
\title{Create script task}
\usage{
task_create_script(
  script,
  chdir = FALSE,
  echo = TRUE,
  environment = "default",
  submit = NULL,
  root = NULL
)
}
\arguments{
\item{script}{Path for the script}

\item{chdir}{Logical, indicating if we should change the working
directory to the directory containing \code{script} before executing
it (similar to the \code{chdir} argument to \code{\link{source}}).}

\item{echo}{Passed through to \code{source} to control printing while
evaluating.  Generally you will want to leave this as \code{TRUE}}

\item{environment}{Name of the hipercow environment to evaluate the
task within.}

\item{submit}{Control over task submission. This will expand over
time once we support specifying resources. The most simple
interface is to use \code{TRUE} here to automatically submit a task,
using your default configuration, or \code{FALSE} to prevent
submission.  The default \code{NULL} will submit a task if a driver
is configured.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
A task id, a string of hex characters. Use this to
interact with the task.
}
\description{
Create a task from a script.  This will arrange to run the file
\code{script} via hipercow.  The script must exist within your hipercow
root, but you may change to the directory of the script as it
executes (otherwise we will evaluate from your current directory
relative to the hipercow root, as usual).
}
