% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_eval}
\alias{task_eval}
\title{Run a task}
\usage{
task_eval(id, envir = .GlobalEnv, verbose = FALSE, root = NULL)
}
\arguments{
\item{id}{The task identifier}

\item{envir}{An environment in which to evaluate the
expression. For non-testing purposes, generally ignore this, the
global environment will be likely the expected environment.}

\item{verbose}{Logical, indicating if we should print information
about what we do as we do it.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
Boolean indicating success (\code{TRUE}) or failure (\code{FALSE})
}
\description{
Run a task that has been created by a \verb{task_create_*} function,
e.g., \code{\link[=task_create_explicit]{task_create_explicit()}}, \code{\link[=task_create_expr]{task_create_expr()}}. Generally
users should not run this function directly.
}
