% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_log_show}
\alias{task_log_show}
\alias{task_log_value}
\title{Get task result}
\usage{
task_log_show(id, root = NULL)

task_log_value(id, root = NULL)
}
\arguments{
\item{id}{The task identifier}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
The value of the queued expression
}
\description{
Get the task log, if the task has produced one.  Tasks run by the
\code{windows} driver will generally produce a log.  A log might be
quite long, and you might want to print it to screen in its
entirity (\code{task_log_show}), return it as character vector
(\code{task_log_value}).
}
