% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_result}
\alias{task_result}
\title{Get task result}
\usage{
task_result(id, follow = TRUE, root = NULL)
}
\arguments{
\item{id}{The task identifier}

\item{follow}{Logical, indicating if we should follow any retried
tasks.}

\item{root}{A hipercow root, or path to it. If \code{NULL} we search up
your directory tree.}
}
\value{
The value of the queued expression
}
\description{
Get the task result. This might be an error if the task has failed.
}
\examples{
hipercow_example_helper()

# Typical usage
id <- task_create_expr(runif(1))
task_wait(id)
task_result(id)

# Tasks that error return error values as results
id <- task_create_expr(readRDS("nosuchfile.rds"))
task_wait(id)
task_result(id)
}
